+++ x/perp/types/keys.go
@@ -6,7 +6,7 @@ var (
 
 	MemStoreKey = "mem_perp"
 
-	// RouterKey is the message route for slashing.
+	// RouterKey is the message route for perp.
 	RouterKey = ModuleName
 
 	// QuerierRoute defines the module's query routing key.